+++ examples/triangle/src/main.rs
@@ -1,0 +1,8 @@
+#[cfg(not(target_arch = "wasm32"))]
+fn main() {
+    sindri_triangle::run();
+}
+
+#[cfg(target_arch = "wasm32")]
+fn main() {}
+